requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
asyncpg==0.30.0
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.20
openai==1.59.9
httpx==0.28.1
python-dotenv==1.0.1
